Move pagination test to seperate file
